wip: initial abstraction of diff rendering
