+++ apps/web/core/hooks/use-local-changes.ts
@@ -1,0 +1,12 @@
+import { useQuery } from '@tanstack/react-query';
+
+import { Change } from '../utils/change';
+
+export const useLocalChanges = (spaceId?: string) => {
+  const { data, isLoading } = useQuery({
+    queryKey: ['changes', spaceId],
+    queryFn: () => Change.fromLocal(spaceId),
+  });
+
+  return [data, isLoading] as const;
+};
